[refactor] 사용자 정보 변경 hook
